+++ boards/quark_d2000_crb/pinmux.c
@@ -149,7 +149,7 @@ static void _quark_d2000_set_mux(uint32_t base, uint32_t pin, uint8_t func)
 }
 
 #ifdef CONFIG_PINMUX_DEV
-static uint32_t pinmux_dev_set(struct device *dev, uint32_t pin, uint8_t func)
+static uint32_t pinmux_dev_set(struct device *dev, uint32_t pin, uint32_t func)
 {
 	struct pinmux_config * const pmux = dev->config->config_info;
 	/*
@@ -184,7 +184,7 @@ static uint32_t pinmux_dev_set(struct device *dev, uint32_t pin, uint8_t func)
 	return DEV_OK;
 }
 
-static uint32_t pinmux_dev_get(struct device *dev, uint32_t pin, uint8_t *func)
+static uint32_t pinmux_dev_get(struct device *dev, uint32_t pin, uint32_t *func)
 {
 	struct pinmux_config * const pmux = dev->config->config_info;
 	/*
@@ -218,7 +218,7 @@ static uint32_t pinmux_dev_get(struct device *dev, uint32_t pin, uint8_t *func)
 	return DEV_OK;
 }
 #else
-static uint32_t pinmux_dev_set(struct device *dev, uint32_t pin, uint8_t func)
+static uint32_t pinmux_dev_set(struct device *dev, uint32_t pin, uint32_t func)
 {
 	ARG_UNUSED(dev);
 	ARG_UNUSED(pin);
@@ -229,7 +229,7 @@ static uint32_t pinmux_dev_set(struct device *dev, uint32_t pin, uint8_t func)
 	return DEV_NOT_CONFIG;
 }
 
-static uint32_t pinmux_dev_get(struct device *dev, uint32_t pin, uint8_t *func)
+static uint32_t pinmux_dev_get(struct device *dev, uint32_t pin, uint32_t *func)
 {
 	ARG_UNUSED(dev);
 	ARG_UNUSED(pin);
